Refactor for pac4j
